[GUI] Remove subfolders and use inheritance for Ui class.
