+++ gui/src/ConfigMain.h
@@ -1,28 +1,42 @@
+//
+// Copyright (C) 2018~2018 by xuzhao9 <i@xuzhao.net>
+//
+// This program is free software; you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation; either version 2 of the License,
+// or (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+// General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with this program; see the file COPYING. If not,
+// see <http://www.gnu.org/licenses/>.
+//
 #ifndef FCITX_RIME_CONFIGMAIN_H
 #define FCITX_RIME_CONFIGMAIN_H
 
 #include <fcitxqtconfiguiwidget.h>
 #include <fcitxqtkeysequencewidget.h>
 
+#include "ui_ConfigMain.h"
 #include "Model.h"
-#include "rime-utils/FcitxRimeConfig.h"
-
-namespace Ui {
-  class MainUI;
-}
+#include "FcitxRimeConfig.h"
 
 namespace fcitx_rime {
-  class ConfigMain : public FcitxQtConfigUIWidget {
+  class ConfigMain : public FcitxQtConfigUIWidget, private Ui::MainUI {
     Q_OBJECT
   public:
     explicit ConfigMain(QWidget* parent = 0);
-    QString title();
+    QString title() override;
     ~ConfigMain();
-    void load();
-    void save();
+    void load() override;
+    void save() override;
     
-    QString addon();
-    QString icon();
+    QString addon() override;
+    QString icon() override;
   public slots:
     void keytoggleChanged();
     void stateChanged();
@@ -33,10 +47,8 @@ namespace fcitx_rime {
     void availIMSelectionChanged();
     void activeIMSelectionChanged();
   private:
-    Ui::MainUI* m_ui;
     FcitxRime* rime;
     FcitxRimeConfigDataModel* model;
-    void test();
     void setFcitxQtKeySeq(char* rime_key, FcitxKeySeq& keyseq);
     void yamlToModel();
     void uiToModel();
@@ -50,7 +62,6 @@ namespace fcitx_rime {
     void setModelFromLayout(QVector<FcitxKeySeq>& model_keys, QLayout *layout);
     void setModelKeysToYaml(QVector<FcitxKeySeq>& model_keys, int type, const char* key);
   };
-  
 }
 
 #endif // FCITX_RIME_CONFIGMAIN_H
