+++ gui/src/Model.h
@@ -1,15 +1,32 @@
+//
+// Copyright (C) 2018~2018 by xuzhao9 <i@xuzhao.net>
+//
+// This program is free software; you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation; either version 2 of the License,
+// or (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+// General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with this program; see the file COPYING. If not,
+// see <http://www.gnu.org/licenses/>.
+//
 #ifndef FCITX_RIME_MODEL_H
 #define FCITX_RIME_MODEL_H
 
 #include <rime_api.h>
 
-#include "fcitx-utils/keysym.h"
-#include "fcitx-utils/keysymgen.h"
-#include <qkeysequence.h>
+#include "keysym.h"
+#include "keysymgen.h"
+#include <QKeySequence>
 #include <QFlags>
 
-#define MAX_SHORTCUTS 3
-#define DEFAULT_PAGE_SIZE 5
+static constexpr int max_shortcuts = 3;
+static constexpr int default_page_size = 5;
 
 namespace fcitx_rime {
   typedef QFlags<fcitx::KeyState> KeyStates;
