+++ gui/src/keynametable.h
@@ -18,9 +18,6 @@
  * License along with this library; see the file COPYING. If not,
  * see <http://www.gnu.org/licenses/>.
  */
-
-
-
 #ifndef _FCITX_UTILS_KEYNAMETABLE_H_
 #define _FCITX_UTILS_KEYNAMETABLE_H_
